feat: show post-event thank you message on home page
Replace the animated activities headline and Add to Calendar button
with a "Thank you!!" message and "Stay tuned for Open Streets Tempe
2027" subhead now that the 2026 event has concluded. The date/time/
location info box stays so visitors can still see when and where it
was. Extract the cycling activities headline into
AnimatedActivities.tsx so it can be dropped back in next year.
